[core] move Qimei to network package
